fix: price Long타입 변경, ItemDto getuuid로 수정

+++ src/main/kotlin/edu/dongguk/nusinsa/dto/ItemDto.kt
@@ -3,7 +3,7 @@ package edu.dongguk.nusinsa.dto
 class ItemDto(
     var id: Long?,
     var name: String,
-    var price:Int,
+    var price:Long,
     var thumbnail: String
 ) {
 }
